pulled maps into their own file so I can work on maps and collision function independently.
