+++ maps.h
@@ -1,0 +1,11 @@
+#ifndef MAPS
+#define MAPS
+
+#include <stdint.h>
+
+void map1();
+
+void map1_collide(int16_t x0, int16_t x1, int16_t x2, int16_t x3, int16_t y0,
+                  int16_t y1, int16_t y2, int16_t y3);
+
+#endif /* MAPS */
